chore: move folders inside /app
