+++ app/_layout.tsx
@@ -30,7 +30,7 @@ SplashScreen.preventAutoHideAsync()
 
 export default function RootLayout() {
   const [loaded, error] = useFonts({
-    SpaceMono: require('../assets/fonts/SpaceMono-Regular.ttf'),
+    SpaceMono: require('./styles/assets/fonts/SpaceMono-Regular.ttf'),
     ...FontAwesome.font
   })
