+++ app/components/pair-details/ConnectionStatus.tsx
