+++ app/components/pair-details/CurrencyInfo.tsx
