+++ app/components/pair-details/PriceCard.tsx
