+++ app/components/pair-details/TechnicalDetails.tsx
