+++ app/components/watchlist/AnimatedWatchlistItem.tsx
