+++ app/components/watchlist/SwipeableWatchlistItem.tsx
