+++ app/components/watchlist/WatchlistItem.tsx
