+++ app/hooks/adapters/useForexStreamAdapter.ts
