+++ app/hooks/ui/useEmptyStateAnimation.ts
