+++ app/hooks/usePairNavigation.ts
