+++ app/lib/constants/Forex.ts
