+++ app/lib/services/forex-service.ts
@@ -1,5 +1,5 @@
 import type { ForexRate, ForexServiceResponse } from '@/types/forex'
-import { FOREX_SERVICE_CONFIG } from '@/constants/Config'
+import { FOREX_SERVICE_CONFIG } from '@/lib/constants/Config'
 import { formatForexResponse, formatPairForApi } from '@/lib/utils/forex-utils'
 
 const FOREX_SERVICE_URL = FOREX_SERVICE_CONFIG.BASE_URL
