+++ app/lib/types/forex.ts
