+++ app/lib/types/forexStore.ts
