+++ app/lib/utils/forex-store-utils.ts
