+++ app/lib/utils/forex-utils.ts
