+++ app/lib/utils/haptics.ts
