+++ app/styles/assets/fonts/SpaceMono-Regular.ttf
