+++ app/styles/assets/images/adaptive-icon.png
