+++ app/styles/assets/images/splash-icon.png
