+++ app/styles/mainScreen.ts
