+++ app/styles/pairDetails.ts
