+++ app/styles/swipeableWatchlistItem.ts
