+++ app/styles/watchlistItem.ts
